seperating Score.java (#161)
* Create ScoreService.java, step 1 in seperating Score.java

* Update Score.java seperating Score and Score Service
